old-tests/submission/Infraware/Forms/contents/Forms/Not Use/input_src.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_src</h3>
	</p>
	
	<li>Is shown image button.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='image' id='input_image'></p>
	</form>

	<script>
		
		var input_image = document.getElementById("input_image");
		input_image.src = "resources/logo.png";

		if (typeof(input_image.src) == "string") {
			var source = input_image.src.substring(input_image.src.length-18, input_image.src.length);
			if (source != "resources/logo.png") {
				test(function() {
					assert_true(false, "src attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("src attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
